vectrix

Struct Matrix

pub struct Matrix<T, const M: usize, const N: usize> { /* private fields */ }

Represents a matrix with constant M rows and constant N columns.

The underlying data is represented as an array and is always stored in column-major order.

See the crate root for usage examples.

Implementations

impl<T> Matrix<T, 1, 1>

pub const fn new(x: T) -> Self

Creates a new vector from the given components.

impl<T> Matrix<T, 1, 2>

pub const fn new(x: T, y: T) -> Self

Creates a new vector from the given components.

impl<T> Matrix<T, 1, 3>

pub const fn new(x: T, y: T, z: T) -> Self

Creates a new vector from the given components.

impl<T> Matrix<T, 1, 4>

pub const fn new(x: T, y: T, z: T, w: T) -> Self

Creates a new vector from the given components.

impl<T> Matrix<T, 1, 5>

pub const fn new(x: T, y: T, z: T, w: T, a: T) -> Self

Creates a new vector from the given components.

impl<T> Matrix<T, 1, 6>

pub const fn new(x: T, y: T, z: T, w: T, a: T, b: T) -> Self

Creates a new vector from the given components.

impl<T> Matrix<T, 2, 1>

pub const fn new(x: T, y: T) -> Self

Creates a new vector from the given components.

impl<T> Matrix<T, 3, 1>

pub const fn new(x: T, y: T, z: T) -> Self

Creates a new vector from the given components.

impl<T> Matrix<T, 4, 1>

pub const fn new(x: T, y: T, z: T, w: T) -> Self

Creates a new vector from the given components.

impl<T> Matrix<T, 5, 1>

pub const fn new(x: T, y: T, z: T, w: T, a: T) -> Self

Creates a new vector from the given components.

impl<T> Matrix<T, 6, 1>

pub const fn new(x: T, y: T, z: T, w: T, a: T, b: T) -> Self

Creates a new vector from the given components.

impl<T, const M: usize, const N: usize> Matrix<T, M, N>

pub fn zero() -> Self
where T: Copy + Zero,

Returns a zero matrix.

pub fn repeat(element: T) -> Self
where T: Copy,

Create a new matrix filled with the given element.

pub fn repeat_with<F>(f: F) -> Self
where F: FnMut() -> T,

Create a new matrix filled with computed elements.

Elements will be filled in column-major order.

pub fn as_slice(&self) -> &[T]

Views the underlying data as a contiguous slice.

pub fn as_mut_slice(&mut self) -> &mut [T]

Views the underlying data as a contiguous mutable slice.

pub fn get<I>(&self, i: I) -> Option<&I::Output>
where I: MatrixIndex<Self>,

Returns a reference to an element in the matrix or None if out of bounds.

pub fn get_mut<I>(&mut self, i: I) -> Option<&mut I::Output>
where I: MatrixIndex<Self>,

Returns a mutable reference to an element in the matrix or None if out of bounds.

pub unsafe fn get_unchecked<I>(&self, i: I) -> &I::Output
where I: MatrixIndex<Self>,

Returns a reference to an element in the matrix without doing any bounds checking.

Safety

Calling this method with an out-of-bounds index is undefined behavior even if the resulting reference is not used.

pub unsafe fn get_unchecked_mut<I>(&mut self, i: I) -> &mut I::Output
where I: MatrixIndex<Self>,

Returns a mutable reference to an element in the matrix without doing any bounds checking.

Safety

Calling this method with an out-of-bounds index is undefined behavior even if the resulting reference is not used.

pub fn row(&self, i: usize) -> &Row<T, M, N>

Returns a reference to the i-th row of this matrix.

pub fn row_mut(&mut self, i: usize) -> &mut Row<T, M, N>

Returns a mutable reference to the i-th row of this matrix.

pub fn column(&self, i: usize) -> &Column<T, M, N>

Returns a reference to the i-th column of this matrix.

pub fn column_mut(&mut self, i: usize) -> &mut Column<T, M, N>

Returns a mutable reference to the i-th column of this matrix.

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator over the underlying data.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Returns a mutable iterator over the underlying data.

pub fn iter_rows(&self) -> IterRows<'_, T, M, N>

Returns an iterator over the rows in this matrix.

pub fn iter_rows_mut(&mut self) -> IterRowsMut<'_, T, M, N>

Returns a mutable iterator over the rows in this matrix.

pub fn iter_columns(&self) -> IterColumns<'_, T, M, N>

Returns an iterator over the columns in this matrix.

pub fn iter_columns_mut(&mut self) -> IterColumnsMut<'_, T, M, N>

Returns a mutable iterator over the columns in this matrix.

pub fn map<F, U>(self, f: F) -> Matrix<U, M, N>
where F: FnMut(T) -> U,

Returns a matrix of the same size as self, with function f applied to each element in column-major order.

pub fn l1_norm(&self) -> T
where T: Copy + Ord + Abs + Zero + Sum<T>,

Returns the L1 norm of the matrix.

Also known as Manhattan Distance or Taxicab norm. L1 Norm is the sum of the magnitudes of the vectors in a space.

Note

If the matrix is a row vector this method might not do what you what you expect. For example:

let row_vector = matrix![1, 2, 3];
assert_eq!(row_vector.l1_norm(), 3);

let column_vector = matrix![1; 2; 3];
assert_eq!(column_vector.l1_norm(), 6);

impl<T, const N: usize> Matrix<T, N, N>

pub fn identity() -> Self
where T: Copy + One + Zero,

Returns an identity matrix.

pub fn diagonal(&self) -> Vector<T, N>
where T: Copy + Zero,

Returns the diagonal of the matrix.